fixed cons_s to Cons_s and argument types for Fisher(..)

+++ Types.jl
@@ -14,15 +14,17 @@ type Fishers
     x::Array{Float64,2} # location
     H::Array{Float64,1} # harvest count
     S::Array{Int32,1} # make(1)/break(-1) friendships
-    MI::Array{Int32,1} # index of steaming or searching
-    CN::Array{Int32,2} # contact network
+    MI::Array{Int64,1} # index of steaming or searching
+    CN::Array{Int64,2} # contact network
     Dmin::Array{Float64,1} # distance to nearest fish
     DXY::Array{Float64,2} # direction vector (unit)
     VR::Array{Float64,1} # speed
-    JJ::Array{Int32,1} # index of nearest fish
-    KK::Array{Int32,1} # 1/0 harvest index
+    JJ::Array{Int64,1} # index of nearest fish
+    KK::Array{Int64,1} # 1/0 harvest index
     cs::Array{Float64,1} # cumulative harvest
-    Dist::Array{Float64,1} # cumulative distance traveled
+    Dist::Array{Float64,1} # cumulative distance traveled 
+
+
 end
 
 #### Output variable for plotting
